+++ phantom-reactivestreams/src/test/scala/com/websudos/phantom/reactivestreams/suites/BatchSubscriberWhiteboxTest.scala
@@ -1,0 +1,87 @@
+/*
+ * Copyright 2013-2015 Websudos, Limited.
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * - Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * - Redistributions in binary form must reproduce the above copyright
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
+ *
+ * - Explicit consent must be obtained from the copyright owner, Websudos Limited before any redistribution is made.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+package com.websudos.phantom.reactivestreams.suites
+
+import com.websudos.phantom.batch.BatchType
+import com.websudos.phantom.reactivestreams._
+import com.websudos.phantom.testkit.suites.PhantomCassandraConnector
+import org.reactivestreams.tck.SubscriberWhiteboxVerification.{SubscriberPuppet, WhiteboxSubscriberProbe}
+import org.reactivestreams.tck.{SubscriberWhiteboxVerification, TestEnvironment}
+import org.reactivestreams.{Subscriber, Subscription}
+
+class BatchSubscriberWhiteboxTest
+  extends SubscriberWhiteboxVerification[Opera](new TestEnvironment(1500)) with PhantomCassandraConnector with TestImplicits {
+
+  override def createSubscriber(probe: WhiteboxSubscriberProbe[Opera]): Subscriber[Opera] = {
+    new BatchSubscriber[OperaTable, Opera](
+      OperaTable,
+      OperaRequestBuilder,
+      5,
+      2,
+      batchType = BatchType.Unlogged,
+      flushInterval = None,
+      completionFn = () => (),
+      errorFn = _ => ()
+    ) {
+
+      override def onSubscribe(s: Subscription): Unit = {
+        super.onSubscribe(s)
+
+        probe.registerOnSubscribe(new SubscriberPuppet {
+
+          override def triggerRequest(elements: Long): Unit = {
+            s.request(elements)
+          }
+
+          override def signalCancel(): Unit = {
+            s.cancel()
+          }
+        })
+      }
+
+      override def onComplete(): Unit = {
+        super.onComplete()
+        probe.registerOnComplete()
+      }
+
+      override def onError(t: Throwable): Unit = {
+        probe.registerOnError(t)
+      }
+
+      override def onNext(t: Opera): Unit = {
+        super.onNext(t)
+        probe.registerOnNext(t)
+      }
+
+    }
+  }
+
+  override def createElement(element: Int): Opera = OperaData.operas(element)
+}